Separated out application view into another component file; adding triggers for multiple components based on side-bar option selection
