+++ client/src/components/ArchiveView.js
@@ -1,0 +1,3 @@
+export const ArchiveView = () => {
+
+};
